src/test/java/com/levep/example/SubjectsDAOTest.java
package com.levep.example;

import com.levelp.example.*;
import org.junit.After;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.Persistence;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.assertNotEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertSame;

public class SubjectsDAOTest {
    private EntityManagerFactory factory;
    private EntityManager em;
    private UsersDAO users;
    private SubjectsDAO subjects;

    @Before
    public void setup() {
        factory = Persistence.createEntityManagerFactory("TestPersistenceUnit");
        em = factory.createEntityManager();
        users = new UsersDAO(em);
        subjects = new SubjectsDAO(em);
    }

    @After
    public void stop() {
        if (em != null) {
            em.close();
        }
        if (factory != null) {
            factory.close();
        }
    }

    @Test
    public void testCreateSubj() {
        em.getTransaction().begin();

        Engineer engineer = users.createEngineer("e2");
        Subject subject = subjects.createSubject(engineer, SubjKind.BUILDING, "78:0001:0001", "addr1");

        em.getTransaction().commit();
    }

    @Test
    public void testFindEngineersSubjs() {
        em.getTransaction().begin();

        Engineer engineer = users.createEngineer("e3");
        Subject subject = subjects.createSubject(engineer, SubjKind.BUILDING, "79:0001:0001", "addr2");
        engineer.setSubjects(Arrays.asList(subject));

        em.getTransaction().commit();

        em.refresh(engineer);

        List<Subject> found = engineer.getSubjects();
        System.out.println(found);
        assertNotNull(found);
        assertNotEquals(0, found.size());
        assertSame(subject, found.get(0));
    }
}
